requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
colorama==0.4.4
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj3-cloudinary-storage==0.0.6
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-cors-headers==3.12.0
django-filter==21.1
django-ip-restriction==1.1.1
django-redis==5.2.0
django-tenants==3.4.2
django-tenants-celery-beat==0.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
django-notifications-hq
et-xmlfile==1.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
msgpack==1.0.4
novu==1.4.0
openpyxl==3.0.9
packaging==21.3
Pillow==10.0.1
prompt-toolkit==3.0.29
psycopg2==2.9.9
psycopg2-binary==2.9.3
pusher-push-notifications==2.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv
pytz==2022.1
redis==4.3.1
requests==2.31.0
sentry-sdk==1.5.12
service-identity==21.1.0
sib-api-v3-sdk==7.5.0
six==1.16.0
sqlparse==0.4.2
tenant-schemas-celery==2.0.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.1.0
wrapt==1.14.1
xlwt==1.3.0
zope.interface==5.4.0
